:recycle: extract jib-specific logic from jib-common to jib-extension
